Updated code

+++ FlowofEnglish_Backend/src/main/java/com/FlowofEnglish/repository/UserAttemptsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserAttemptsRepository extends JpaRepository<UserAttempts, Integer> {
+public interface UserAttemptsRepository extends JpaRepository<UserAttempts, Long> {
     // Custom query methods can be added here if necessary
 }
